remove getProject
